+++ crates/swc_ecma_parser/tests/typescript/top-level-await/input.ts
@@ -1,0 +1,6 @@
+await 1 + 2
+await 1 * 2
+await 1 || 2
+await null ?? 3
+await a instanceof await b
+await a in await b