+++ crates/swc_ecma_transforms_optimization/src/simplify/branch/tests.rs
@@ -1,4 +1,5 @@
-use swc_common::chain;
+use swc_common::{chain, Mark};
+use swc_ecma_transforms_base::resolver::resolver_with_mark;
 
 use super::{super::expr_simplifier, dead_branch_remover};
 
@@ -6,7 +7,15 @@ macro_rules! test_stmt {
     ($l:expr, $r:expr) => {
         swc_ecma_transforms_testing::test_transform(
             ::swc_ecma_parser::Syntax::default(),
-            |_| chain!(expr_simplifier(Default::default()), dead_branch_remover()),
+            |_| {
+                let top_level_mark = Mark::fresh(Mark::root());
+
+                chain!(
+                    resolver_with_mark(top_level_mark),
+                    expr_simplifier(top_level_mark, Default::default()),
+                    dead_branch_remover(top_level_mark)
+                )
+            },
             $l,
             $r,
             true,